pyproject.toml
[tool.poetry]
name = "repl-nix-iiqexcelparser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_iiqexcelparser"}]

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "^3.1.2"
flask = "^3.0.2"
pyexiv2 = "^2.12.0"
pyexiftool = "^0.5.6"
pandas = "^2.2.1"
werkzeug = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
